+++ src/main/java/net/pwall/util/ImmutableCollection.java
@@ -2,7 +2,7 @@
  * @(#) ImmutableCollection.java
  *
  * immutables  High-performance immutable collections
- * Copyright (c) 2021 Peter Wall
+ * Copyright (c) 2021, 2022 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -107,6 +107,21 @@ public class ImmutableCollection<T> extends ImmutableCollectionBase<T, T> implem
         if (target.length > length)
             target[length] = null;
         return a;
+    }
+
+    /**
+     * Get the element at the specified index.  This is not part of the standard {@link Collection} interface, but it
+     * allows iteration over the members of a collection without needing to instantiate an {@link Iterator}.
+     *
+     * The function does not check the array index because the indexing
+     * operation on the array will do that anyway.
+     *
+     * @param   index       the index
+     * @return              the element at that index
+     * @throws  IndexOutOfBoundsException   if the index is less than 0 or greater than the length of the list
+     */
+    public T get(int index) {
+        return array[index];
     }
 
     /**